Move AWS-specific code to its own file
In another step towards unifying the aws and master branches,
move all the AWS-specific code to its own file and use
a weak symbol to call into the platform-specific code.  This
patch (minus the platform-aws.c file) could be merged into
master and master would function without any AWS-specific hooks.
