+++ src/Makefile.am
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2018-2019, Amazon.com, Inc. or its affiliates. All rights reserved.
+# Copyright (c) 2018-2023, Amazon.com, Inc. or its affiliates. All rights reserved.
 #
 # See LICENSE.txt for license information
 #
@@ -10,10 +10,20 @@ if HAVE_CUDA
   AM_LDFLAGS = -lcudart
 endif HAVE_CUDA
 
+sources = \
+	nccl_ofi_net.c \
+	tracepoint.c \
+	compat_v6.c \
+	compat_v5.c \
+	compat_v4.c \
+	compat_v3.c \
+	compat_v2.c \
+	platform-aws.c
+
 if ENABLE_NEURON
   lib_LTLIBRARIES = libnccom-net.la
-  libnccom_net_la_SOURCES = nccl_ofi_net.c tracepoint.c compat_v6.c compat_v5.c compat_v4.c compat_v3.c compat_v2.c
+  libnccom_net_la_SOURCES = $(sources)
 else
   lib_LTLIBRARIES = libnccl-net.la
-  libnccl_net_la_SOURCES = nccl_ofi_net.c tracepoint.c compat_v6.c compat_v5.c compat_v4.c compat_v3.c compat_v2.c
+  libnccl_net_la_SOURCES = $(sources)
 endif
