+++ src/nccl_ofi_net.c
@@ -23,28 +23,6 @@
 #define EFA_PROVIDER_NAME "efa"
 #define IS_EFA_PROVIDER(NAME) (strcmp((NAME), EFA_PROVIDER_NAME)==0)
 
-struct ec2_platform_data {
-	const char* name;
-	const char* topology;
-	int default_dup_conns;
-} platform_data_map[] = {
-	[0] = {
-		.name = "p4d.24xlarge",
-		.topology = "p4d-24xl-topo.xml",
-		.default_dup_conns = 0
-	},
-	[1] = {
-		.name = "p4de.24xlarge",
-		.topology = "p4de-24xl-topo.xml",
-		.default_dup_conns = 0
-	},
-	[2] = {
-		.name = "p3dn.24xlarge",
-		.topology = NULL,
-		.default_dup_conns = 4
-	},
-};
-
 static uint32_t libversion = 0;
 /* NICs info list for a provider */
 struct fi_info* ofi_info_list = NULL;
@@ -1285,146 +1263,6 @@ static inline ncclResult_t nccl_ofi_progress(nccl_ofi_t *nccl_ofi_comp)
 	return ofi_process_cq(nccl_ofi_comp);
 }
 
-/*
- * @brief	Provides EC2 platform type as reported by the
- * 		first line of
- *		/sys/devices/virtual/dmi/id/product_name.
- *		Users of this API *should* free the buffer when a
- *		Non-NULL string is returned.
- *
- * @return	NULL, on allocation and file system error
- * 		EC2 platform type, on success
- */
-static const char* get_platform_type()
-{
-	char file[] = "/sys/devices/virtual/dmi/id/product_name";
-	FILE *fd = NULL;
-	char ch;
-	size_t len = 0;
-	size_t platform_type_len = 64;
-	char *platform_type = NULL;
-
-	fd = fopen(file, "r");
-	if (fd == NULL) {
-		NCCL_OFI_WARN("Error opening file: %s", file);
-		goto error;
-	}
-
-	platform_type = (char *)malloc(sizeof(char)*platform_type_len);
-	if (platform_type == NULL) {
-		NCCL_OFI_WARN("Unable to allocate platform type");
-		goto error;
-	}
-
-	/* Read first line of the file, reallocing the buffer as necessary */
-	while ((feof(fd) == 0) && (ferror(fd) == 0) && ((ch = fgetc(fd)) != '\n')) {
-		platform_type[len++] = ch;
-		if (len >= platform_type_len) {
-			platform_type = realloc(platform_type, len + platform_type_len);
-		}
-	}
-
-	if (ferror(fd)) {
-		NCCL_OFI_WARN("Error reading file: %s", file);
-		goto error;
-	}
-
-	platform_type[len] = '\0';
-
-	NCCL_OFI_TRACE(NCCL_INIT | NCCL_NET, "Read %d bytes. EC2 platform type is %s", len, platform_type);
-
-	fclose(fd);
-	return platform_type;
-
-error:
-	if (platform_type)
-		free(platform_type);
-	if (fd)
-		fclose(fd);
-	return platform_type;
-}
-
-/*
- * @brief	Returns AWS-specific platform data, if found
- *
- * @input	Platform type
- *
- * @return	NULL, if no topology found
- * 		Topology structure, if match found
- */
-struct ec2_platform_data *get_platform_data(const char *platform_type)
-{
-	const size_t platform_n = sizeof(platform_data_map)/sizeof(platform_data_map[0]);
-
-	for (size_t idx = 0; idx < platform_n; idx++) {
-		if (strcmp(platform_type, platform_data_map[idx].name) == 0)
-			return &platform_data_map[idx];
-	}
-
-	return NULL;
-}
-
-/*
- * @brief	Update NCCL's system topology using static pre-configured topology
- * 		files for supported EC2 platform types.
- *
- * @return	0, when we are succesfully able to update NCCL topology or
- * 		   if we find no match
- * 		error, on failure
- */
-static ncclResult_t aws_platform_init(void)
-{
-	int ret = ncclSuccess;
-	int rc = 0;
-	struct ec2_platform_data *platform_data;
-
-	const char *platform_type = get_platform_type();
-	if (platform_type == NULL) {
-		ret = ncclSystemError;
-		goto exit;
-	}
-
-	platform_data = get_platform_data(platform_type);
-	if (!platform_data) goto exit;
-
-	const char *topo_file = platform_data->topology;
-	if (topo_file == NULL) {
-		/* No topology file exists for the given platform so return success */
-		goto exit;
-	} else {
-		/* Update topology */
-		char topology_path[PATH_MAX];
-
-		rc = snprintf(topology_path, sizeof(topology_path), "%s/%s",
-				XML_DIR, topo_file);
-		if (rc < 0 || rc >= sizeof(topology_path)) {
-			NCCL_OFI_WARN("Error occurred while forming the complete topology XML file path. RC: %d, Buffer Size: %d, XML dir: %s, Topology file: %s",
-					rc, PATH_MAX, XML_DIR, topo_file);
-			ret = ncclSystemError;
-			goto exit;
-		}
-
-		NCCL_OFI_INFO(NCCL_INIT | NCCL_NET,
-				"Running on %s platform, Setting NCCL_TOPO_FILE environment variable to %s",
-				platform_type, topology_path);
-
-		rc = setenv("NCCL_TOPO_FILE", topology_path, 1);
-		if (rc != 0) {
-			NCCL_OFI_WARN("Unable to set NCCL_TOPO_FILE");
-			ret = ncclSystemError;
-			goto exit;
-		}
-	}
-
-	if (nic_dup_conns == 0)
-		nic_dup_conns = platform_data->default_dup_conns;
-
-exit:
-	if (platform_type)
-		free((char *)platform_type);
-	return ret;
-}
-
 ncclResult_t nccl_net_ofi_init(ncclDebugLogger_t logFunction)
 {
 	ncclResult_t ret = ncclSuccess;
@@ -1449,21 +1287,11 @@ ncclResult_t nccl_net_ofi_init(ncclDebugLogger_t logFunction)
 
 	nic_dup_conns = ofi_nccl_nic_dup_conns();
 
-	/*
-	 * Use a static pre-configured topology for supported EC2 platform types
-	 *
-	 * For example, for P4D platform, while the actual physical topology is
-	 * 2 GPUs and a NIC behind a PCIe switch, the AWS hypervisor presents
-	 * this as a separate PCI bus without switch. To enable GPUDirect,
-	 * plugin silently overrides the system topology using `NCCL_TOPO_FILE`
-	 * environment variable to inform NCCL of the underlying hardware topology.
-	 *
-	 * This topology information helps NCCL to form optimal
-	 * graphs by using right GPU-NIC pairs for transfers through network.
-	 */
-	ret = aws_platform_init();
-	if (ret != ncclSuccess)
-		goto exit;
+	if (platform_init) {
+		ret = platform_init();
+		if (ret != ncclSuccess)
+			goto exit;
+	}
 
 	/*
 	 * FI_EFA_FORK_SAFE environment variable tells Libfabric to enable
